Creating Landing Page (#1)
* Update git ignore

* update the requirements.txt

* Creating flask package structure

* Fix data update issue in Flask app

The website now launches correctly, but there was an issue with updating the data in the Excel file. Users can now register with their name and phone number, and the data is correctly added to the Excel file. However, further testing is still needed to ensure that the app functions correctly under all conditions.

* update index.html in Flask app

Update landing page and fix Excel file issue

- Updated landing page HTML to display in Hebrew
- Fixed issue with Excel file creation that was causing a new file to be created each time the app was executed
- Users can now register with their name and phone number, and the data is correctly added to the existing Excel file

Further testing is still needed to ensure that the app functions correctly under all conditions.

* -Rename index.html to landing_page.html
-Add maintenance mode functionality
-Create landing page and registration form
-Implement register route to store student data
-Add configuration to app for maintenance mode
-Refactor app initialization to use configuration file

* -Add Timestemp for __init__.py

* -update the look of landing_page.html

* -update the title of the base.html

* -update the landing_page.html
* make it look beter

* - Adding notes to the line for better documentation.

* - Move all HTML to views directory

* - Refactored into a modular structure using blueprints.

* - Deleted these files
Right now, I don't see any use of them

* - Refactored the __init__.py file
  Use the 'create_flask_app' function to create and run the Flask application.

* -Moved the static folder from views to the root directory

* -Updated the static folder path in the main_bp blueprint

* -Added a new 'create_flask_app' function to the __init__.py file to create the Flask application.

* -Update redirect function call to use the full endpoint path for the landing page view function.

* -feat: Add flash message to registration form in landing page
Updated the HTML code in landing_page.html to include a flash message below the registration form. The flash message will only appear once the user has submitted the registration form.

* -update flash message

* -Add init_test.py
This is for additional setup or configurations

* -Add unit tests for landing page and registration route

Created a new file test_routes.py inside the tests/ directory, which includes two unit tests for the landing page and registration route. The setUp method sets up the app for testing, and the test_landing_page and test_register methods test the respective routes. The tests are executed using the unittest module, and the if __name__ == '__main__' block allows the tests to be run with the python test_routes.py command.

* -Moved to landing_page

* -Add Dockerfile for landing page application

* -Add Jenkinsfile

* -update gitignore

* -Refactored Dockerfile for landing_page Flask application

* -Update README.md

* -Moved requirements.txt from landing_page

* -Add unit_tests.py for testing the routes of our application

* -Add unit_tests.py for testing the routes of our application